refactor home screen components
