+++ app/(tabs)/index.tsx
@@ -1,11 +1,10 @@
-import React, { useState, useEffect, useRef, Suspense, lazy } from 'react';
+import React, { useState, useEffect, Suspense, lazy } from 'react';
 import {
   View,
   Text,
   StyleSheet,
   ScrollView,
   TouchableOpacity,
-  Dimensions,
   Modal,
   RefreshControl,
   useWindowDimensions,
@@ -13,14 +12,18 @@ import {
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { useLocalSearchParams } from 'expo-router';
 import useAppRouter from 'hooks/useAppRouter';
-import { Plus, Pencil, X, ArrowUpDown } from 'lucide-react-native';
+import { Plus, X, ArrowUpDown } from 'lucide-react-native';
 import { Product, Category, HeroBanner } from '@/types';
 import { useHome } from '@/features/home/hooks/useHome';
 import { useAuth } from '@/features/auth/AuthContext';
 import { useLanguage } from '@/contexts/LanguageContext';
 import { useTheme } from '@/contexts/ThemeContext';
 import HomeHeader from '@/features/home/components/HomeHeader';
+import SearchBar from '@/features/home/components/SearchBar';
+import PriceRange from '@/features/home/components/PriceRange';
 import CategoryTabs from '@/features/home/components/CategoryTabs';
+import CTABecomeSeller from '@/features/home/components/CTABecomeSeller';
+import BannerArea from '@/features/home/components/BannerArea';
 const ProductGrid = lazy(() => import('@/features/home/components/ProductGrid'));
 import Spinner from '@/shared/ui/Spinner';
 import EmptyState from '@/shared/ui/EmptyState';
@@ -28,13 +31,9 @@ import ErrorBoundary from '@/components/ui/ErrorBoundary';
 const BannerFormModal = lazy(() => import('@/components/BannerFormModal'));
 const CartModal = lazy(() => import('@/features/cart/components/CartModal'));
 const ProductFormModal = lazy(() => import('@/features/products/components/ProductFormModal'));
-const SmartImage = lazy(() => import('@/components/SmartImage'));
 const InfoModal = lazy(() => import('@/components/InfoModal'));
 import { useHomeScreen, SortOption } from '@/features/home/hooks/useHomeScreen';
 
-const { width } = Dimensions.get('window');
-const BANNER_WIDTH = width - 32;
-const BANNER_HEIGHT = (BANNER_WIDTH * 9) / 16;
 
 function HomeScreenContent() {
   const { push } = useAppRouter();
@@ -52,7 +51,6 @@ function HomeScreenContent() {
     removeProduct,
     error,
   } = useHome();
-  const [currentBannerIndex, setCurrentBannerIndex] = useState(0);
   const [bannerFormVisible, setBannerFormVisible] = useState(false);
   const [editingBanner, setEditingBanner] = useState<HeroBanner | null>(null);
   const [productFormVisible, setProductFormVisible] = useState(false);
@@ -68,7 +66,6 @@ function HomeScreenContent() {
   const { isStoreOwner } = useAuth();
   const { t } = useLanguage();
   const { colors } = useTheme();
-  const bannerScrollRef = useRef<ScrollView>(null);
 
   const {
     filteredProducts,
@@ -106,24 +103,6 @@ function HomeScreenContent() {
       setShowCartModal(true);
     }
   }, [params.showCart]);
-
-  useEffect(() => {
-    // Auto-scroll banner for web compatibility
-    if (heroBanners.length > 1) {
-      const bannerInterval = setInterval(() => {
-        setCurrentBannerIndex((prevIndex) => {
-          const nextIndex = (prevIndex + 1) % heroBanners.length;
-          bannerScrollRef.current?.scrollTo({
-            x: nextIndex * (width - 32),
-            animated: true,
-          });
-          return nextIndex;
-        });
-      }, 5000);
-
-      return () => clearInterval(bannerInterval);
-    }
-  }, [heroBanners.length]);
 
   useEffect(() => {
     if (error) {
@@ -212,68 +191,6 @@ function HomeScreenContent() {
     </TouchableOpacity>
   );
 
-  const renderBanner = (item: HeroBanner, index: number) => (
-    <View key={item.id} style={styles.heroBanner}>
-      <TouchableOpacity
-        style={styles.bannerTouchable}
-        onPress={() => push(`/category/${item.category}`)}
-      >
-        <Suspense fallback={<Spinner />}>
-          <SmartImage
-            uri={item.image}
-            width={BANNER_WIDTH}
-            height={BANNER_HEIGHT}
-            contentFit="cover"
-          />
-        </Suspense>
-        <View
-          pointerEvents="none"
-          style={[
-            styles.heroOverlay,
-            { backgroundColor: colors.background + '66' },
-          ]}
-        >
-          <View style={styles.heroContent}>
-            {item.discount ? (
-              <Text
-                style={[
-                  styles.heroDiscount,
-                  {
-                    color: colors.text.inverse,
-                    backgroundColor: colors.gold,
-                  },
-                ]}
-              >
-                {item.discount} הנחה
-              </Text>
-            ) : null}
-            <Text style={[styles.heroTitle, { color: colors.gold }]}>
-              {item.title}
-            </Text>
-
-            <Text style={[styles.heroSubtitle, { color: colors.gold }]}>
-              {item.subtitle}
-            </Text>
-          </View>
-        </View>
-      </TouchableOpacity>
-
-      {isStoreOwner && (
-        <View style={styles.bannerAdminActions}>
-          <TouchableOpacity
-            style={[
-              styles.bannerAdminButton,
-              { backgroundColor: colors.background + 'CC' },
-            ]}
-            onPress={() => editBanner(item)}
-          >
-            <Pencil size={16} color="#FFFFFF" />
-          </TouchableOpacity>
-        </View>
-      )}
-    </View>
-  );
-
   // Function to determine product item width based on screen size
   const getProductItemWidth = () => {
     if (windowWidth >= 1024) {
@@ -311,7 +228,8 @@ function HomeScreenContent() {
     <SafeAreaView
       style={[styles.container, { backgroundColor: colors.background }]}
     >
-    <HomeHeader
+    <HomeHeader />
+    <SearchBar
       searchQuery={searchQuery}
       onSearchChange={setSearchQuery}
     />
@@ -326,87 +244,20 @@ function HomeScreenContent() {
           categories={categories}
           selectedCategory={selectedCategory}
           setSelectedCategory={setSelectedCategory}
+        />
+      <PriceRange
           minPrice={minPrice}
           setMinPrice={setMinPrice}
           maxPrice={maxPrice}
           setMaxPrice={setMaxPrice}
         />
-        <TouchableOpacity
-          style={[styles.becomeSellerButton, { backgroundColor: colors.gold }]}
-          onPress={() => push('/stores/create')}
-          accessibilityRole="link"
-        >
-          <Text style={[styles.becomeSellerText, { color: colors.text.inverse }]}>Become a Seller</Text>
-        </TouchableOpacity>
-        {/* Hero Banner Carousel */}
-        <View style={styles.bannerContainer}>
-          <View style={styles.bannerHeader}>
-            {isStoreOwner && (
-              <TouchableOpacity
-                style={[
-                  styles.addBannerButton,
-                  { backgroundColor: colors.gold },
-                ]}
-                onPress={addBanner}
-              >
-                <Plus size={20} color={colors.text.inverse} />
-                <Text
-                  style={[styles.addBannerText, { color: colors.text.inverse }]}
-                >
-                  {t('banner.addBanner')}
-                </Text>
-              </TouchableOpacity>
-            )}
-          </View>
-
-          {heroBanners.length > 0 ? (
-            <>
-              <ScrollView
-                ref={bannerScrollRef}
-                horizontal
-                pagingEnabled
-                showsHorizontalScrollIndicator={false}
-                onMomentumScrollEnd={(event) => {
-                  const newIndex = Math.round(
-                    event.nativeEvent.contentOffset.x / (width - 32)
-                  );
-                  setCurrentBannerIndex(newIndex);
-                }}
-                contentContainerStyle={styles.bannerScrollContent}
-              >
-                {heroBanners.map((item, index) => renderBanner(item, index))}
-              </ScrollView>
-
-              {/* Banner Indicators */}
-              {heroBanners.length > 1 && (
-                <View pointerEvents="none" style={styles.bannerIndicators}>
-                  {heroBanners.map((_, index) => (
-                    <View
-                      key={index}
-                      style={[
-                        styles.indicator,
-                        index === currentBannerIndex && [
-                          styles.activeIndicator,
-                          { backgroundColor: colors.gold },
-                        ],
-                      ]}
-                    />
-                  ))}
-                </View>
-              )}
-            </>
-          ) : (
-            <EmptyState
-              icon={Plus}
-              title={t('home.noBanners')}
-              message={
-                isStoreOwner ? t('home.addBanners') : t('home.bannersComingSoon')
-              }
-              actionText={isStoreOwner ? t('banner.addBanner') : undefined}
-              onAction={isStoreOwner ? addBanner : undefined}
-            />
-          )}
-        </View>
+      <CTABecomeSeller />
+      <BannerArea
+          heroBanners={heroBanners}
+          isStoreOwner={isStoreOwner}
+          onAddBanner={addBanner}
+          onEditBanner={editBanner}
+        />
 
         {/* Categories Section */}
         <View style={styles.section}>
@@ -630,121 +481,6 @@ export default function HomeScreen() {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-  },
-  becomeSellerButton: {
-    borderRadius: 12,
-    paddingVertical: 12,
-    paddingHorizontal: 16,
-    alignItems: 'center',
-    marginHorizontal: 16,
-    marginBottom: 24,
-  },
-  becomeSellerText: {
-    fontSize: 16,
-    fontWeight: '600',
-  },
-  bannerContainer: {
-    height: BANNER_HEIGHT + 40,
-    marginBottom: 24,
-  },
-  bannerHeader: {
-    flexDirection: 'row',
-    justifyContent: 'flex-start',
-    paddingHorizontal: 16,
-    marginBottom: 8,
-  },
-  addBannerButton: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    paddingHorizontal: 12,
-    paddingVertical: 6,
-    borderRadius: 16,
-  },
-  addBannerText: {
-    fontSize: 12,
-    fontWeight: '600',
-    marginLeft: 4,
-  },
-  bannerScrollContent: {
-    paddingHorizontal: 16,
-  },
-  heroBanner: {
-    width: BANNER_WIDTH,
-    height: BANNER_HEIGHT,
-    marginRight: 16,
-    borderRadius: 16,
-    overflow: 'hidden',
-    position: 'relative',
-  },
-  bannerTouchable: {
-    width: '100%',
-    height: '100%',
-  },
-  heroOverlay: {
-    position: 'absolute',
-    top: 0,
-    start: 0,
-    end: 0,
-    bottom: 0,
-    justifyContent: 'center',
-    zIndex: 0,
-  },
-  heroContent: {
-    paddingHorizontal: 20,
-    alignItems: 'flex-end',
-  },
-  heroDiscount: {
-    fontSize: 16,
-    fontWeight: 'bold',
-    paddingHorizontal: 12,
-    paddingVertical: 4,
-    borderRadius: 12,
-    alignSelf: 'flex-start',
-    marginBottom: 8,
-  },
-  heroTitle: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 4,
-    textAlign: 'end',
-  },
-  heroSubtitle: {
-    fontSize: 16,
-    marginBottom: 16,
-    textAlign: 'end',
-  },
-  bannerAdminActions: {
-    position: 'absolute',
-    top: 8,
-    start: 8,
-    flexDirection: 'row',
-  },
-  bannerAdminButton: {
-    borderRadius: 16,
-    width: 32,
-    height: 32,
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginRight: 4,
-  },
-  bannerIndicators: {
-    position: 'absolute',
-    bottom: 16,
-    start: 0,
-    end: 0,
-    flexDirection: 'row',
-    justifyContent: 'center',
-    zIndex: 0,
-  },
-  indicator: {
-    width: 8,
-    height: 8,
-    borderRadius: 4,
-    backgroundColor: 'rgba(255,255,255,0.5)',
-    marginHorizontal: 4,
-  },
-  activeIndicator: {
-    width: 16,
   },
   section: {
     paddingHorizontal: 16,
