+++ src/features/home/components/BannerArea.tsx
@@ -1,0 +1,243 @@
+import React, { useState, useEffect, useRef, Suspense, lazy } from 'react';
+import { View, Text, ScrollView, TouchableOpacity, Dimensions, StyleSheet } from 'react-native';
+import { Plus, Pencil } from 'lucide-react-native';
+import { HeroBanner } from '@/types';
+import { useTheme } from '@/contexts/ThemeContext';
+import { useLanguage } from '@/contexts/LanguageContext';
+import useAppRouter from 'hooks/useAppRouter';
+import EmptyState from '@/shared/ui/EmptyState';
+import Spinner from '@/shared/ui/Spinner';
+
+const SmartImage = lazy(() => import('@/components/SmartImage'));
+
+const { width } = Dimensions.get('window');
+const BANNER_WIDTH = width - 32;
+const BANNER_HEIGHT = (BANNER_WIDTH * 9) / 16;
+
+interface BannerAreaProps {
+  heroBanners: HeroBanner[];
+  isStoreOwner: boolean;
+  onAddBanner: () => void;
+  onEditBanner: (banner: HeroBanner) => void;
+}
+
+export default function BannerArea({ heroBanners, isStoreOwner, onAddBanner, onEditBanner }: BannerAreaProps) {
+  const { colors } = useTheme();
+  const { t } = useLanguage();
+  const { push } = useAppRouter();
+  const bannerScrollRef = useRef<ScrollView>(null);
+  const [currentIndex, setCurrentIndex] = useState(0);
+
+  useEffect(() => {
+    if (heroBanners.length > 1) {
+      const bannerInterval = setInterval(() => {
+        setCurrentIndex((prevIndex) => {
+          const nextIndex = (prevIndex + 1) % heroBanners.length;
+          bannerScrollRef.current?.scrollTo({ x: nextIndex * (width - 32), animated: true });
+          return nextIndex;
+        });
+      }, 5000);
+      return () => clearInterval(bannerInterval);
+    }
+  }, [heroBanners.length]);
+
+  const renderBanner = (item: HeroBanner) => (
+    <View key={item.id} style={styles.heroBanner}>
+      <TouchableOpacity
+        style={styles.bannerTouchable}
+        onPress={() => push(`/category/${item.category}`)}
+      >
+        <Suspense fallback={<Spinner />}>
+          <SmartImage
+            uri={item.image}
+            width={BANNER_WIDTH}
+            height={BANNER_HEIGHT}
+            contentFit="cover"
+          />
+        </Suspense>
+        <View
+          pointerEvents="none"
+          style={[
+            styles.heroOverlay,
+            { backgroundColor: colors.background + '66' },
+          ]}
+        >
+          <View style={styles.heroContent}>
+            {item.discount ? (
+              <Text
+                style={[
+                  styles.heroDiscount,
+                  {
+                    color: colors.text.inverse,
+                    backgroundColor: colors.gold,
+                  },
+                ]}
+              >
+                {item.discount} הנחה
+              </Text>
+            ) : null}
+            <Text style={[styles.heroTitle, { color: colors.gold }]}>
+              {item.title}
+            </Text>
+
+            <Text style={[styles.heroSubtitle, { color: colors.gold }]}>
+              {item.subtitle}
+            </Text>
+          </View>
+        </View>
+      </TouchableOpacity>
+
+      {isStoreOwner && (
+        <View style={styles.bannerAdminActions}>
+          <TouchableOpacity
+            style={[
+              styles.bannerAdminButton,
+              { backgroundColor: colors.background + 'CC' },
+            ]}
+            onPress={() => onEditBanner(item)}
+          >
+            <Pencil size={16} color="#FFFFFF" />
+          </TouchableOpacity>
+        </View>
+      )}
+    </View>
+  );
+
+  return (
+    <View style={styles.container}>
+      <View style={styles.bannerHeader}>
+        {isStoreOwner && (
+          <TouchableOpacity
+            style={[styles.addBannerButton, { backgroundColor: colors.gold }]}
+            onPress={onAddBanner}
+          >
+            <Plus size={20} color={colors.text.inverse} />
+            <Text style={[styles.addBannerText, { color: colors.text.inverse }]}>
+              {t('banner.addBanner')}
+            </Text>
+          </TouchableOpacity>
+        )}
+      </View>
+
+      {heroBanners.length > 0 ? (
+        <>
+          <ScrollView
+            ref={bannerScrollRef}
+            horizontal
+            pagingEnabled
+            showsHorizontalScrollIndicator={false}
+            onMomentumScrollEnd={(event) => {
+              const newIndex = Math.round(
+                event.nativeEvent.contentOffset.x / (width - 32)
+              );
+              setCurrentIndex(newIndex);
+            }}
+            contentContainerStyle={styles.bannerScrollContent}
+          >
+            {heroBanners.map((item) => renderBanner(item))}
+          </ScrollView>
+
+          {heroBanners.length > 1 && (
+            <View pointerEvents="none" style={styles.bannerIndicators}>
+              {heroBanners.map((_, index) => (
+                <View
+                  key={index}
+                  style={[
+                    styles.indicator,
+                    index === currentIndex && [
+                      styles.activeIndicator,
+                      { backgroundColor: colors.gold },
+                    ],
+                  ]}
+                />
+              ))}
+            </View>
+          )}
+        </>
+      ) : (
+        <EmptyState
+          icon={Plus}
+          title={t('home.noBanners')}
+          message={
+            isStoreOwner ? t('home.addBanners') : t('home.bannersComingSoon')
+          }
+          actionText={isStoreOwner ? t('banner.addBanner') : undefined}
+          onAction={isStoreOwner ? onAddBanner : undefined}
+        />
+      )}
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: { marginBottom: 24 },
+  bannerHeader: {
+    flexDirection: 'row',
+    justifyContent: 'flex-end',
+    marginHorizontal: 16,
+    marginBottom: 8,
+  },
+  addBannerButton: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingVertical: 8,
+    paddingHorizontal: 12,
+    borderRadius: 12,
+  },
+  addBannerText: {
+    marginStart: 8,
+    fontWeight: '600',
+  },
+  bannerScrollContent: { paddingHorizontal: 16 },
+  heroBanner: { marginRight: 16, position: 'relative' },
+  bannerTouchable: { borderRadius: 12, overflow: 'hidden' },
+  heroOverlay: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+    justifyContent: 'flex-end',
+    borderRadius: 12,
+  },
+  heroContent: { padding: 16 },
+  heroDiscount: {
+    fontSize: 12,
+    fontWeight: '600',
+    paddingHorizontal: 8,
+    paddingVertical: 4,
+    borderRadius: 8,
+    alignSelf: 'flex-start',
+    marginBottom: 8,
+  },
+  heroTitle: { fontSize: 18, fontWeight: '700' },
+  heroSubtitle: { fontSize: 14, marginTop: 4 },
+  bannerIndicators: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+    marginTop: 8,
+  },
+  indicator: {
+    width: 8,
+    height: 8,
+    borderRadius: 4,
+    backgroundColor: '#ccc',
+    marginHorizontal: 4,
+  },
+  activeIndicator: {},
+  bannerAdminActions: {
+    position: 'absolute',
+    top: 8,
+    right: 8,
+    flexDirection: 'row',
+    gap: 8,
+  },
+  bannerAdminButton: {
+    width: 32,
+    height: 32,
+    borderRadius: 16,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
+
