+++ src/features/home/components/HomeHeader.tsx
@@ -1,18 +1,7 @@
 import React from 'react';
 import GlobalHeader from '@/components/GlobalHeader';
 
-interface HomeHeaderProps {
-  searchQuery: string;
-  onSearchChange: (value: string) => void;
-}
-
-export default function HomeHeader({ searchQuery, onSearchChange }: HomeHeaderProps) {
-  return (
-    <GlobalHeader
-      searchQuery={searchQuery}
-      onSearchChange={onSearchChange}
-      showSearch
-    />
-  );
+export default function HomeHeader() {
+  return <GlobalHeader showSearch={false} />;
 }
